[ gud-mi.el ] Fixed: gud-watch using delted completion function
After checking the file history I don't actually understand why
gud-gdb-complete-command was ever used for this file because it hasn't
been working for a *very very* long time. Something like 15 years. It
wouldn't be caught because the bug is only invoked when calling
gud-watch with a prefix argument

It appears the idea was it would insert provide completion results for
user but that doesn't make a huge amount of sense because gdb only
provides the ability to trivially list local variables, which is a
very small subset of a program can actually use. So I just replaced it
in favour of a blank expression input